Move FrameRateManager to separate file
